051-kafka-spark-streaming/requirements.txt
kafka-python==2.0.2
mypy==0.961
mypy-extensions==0.4.3
py4j==0.10.9.5
pyspark==3.3.0
tomli==2.0.1
typing_extensions==4.3.0
